+++ packages/core/modules/utils/jsonLogicUtils.js
@@ -14,16 +14,48 @@ export const customJsonLogicOperations = {
   JSX: (type, props) => ({type, props}),
   mergeObjects: (obj1, obj2) => ({...obj1, ...obj2}),
   fromEntries: (entries) => Object.fromEntries(entries),
+  //
+  // string
+  //
+  toLowerCase: (str) => str.toLowerCase(),
+  toUpperCase: (str) => str.toUpperCase(),
   strlen: (str) => (str?.length || 0),
   regexTest: (str, pattern, flags) => str?.match(new RegExp(pattern, flags)) != null,
+  //
+  // date / datetime
+  //
+  "date==": (a, b) => {
+    if (a == null || b == null) {
+      return false;
+    }
+    const dateA = moment(a).startOf("day");
+    const dateB = moment(b).startOf("day");
+    return dateA.isSame(dateB); 
+  },
+  "date!=": (a, b) => { return !customJsonLogicOperations["date=="](a, b); },
+  "datetime==": (a, b) => {
+    if (a == null || b == null) {
+      return false;
+    }
+    const dateA = moment(a);
+    const dateB = moment(b);
+    return dateA.isSame(dateB); 
+  },
+  "datetime!=": (a, b) => { return !customJsonLogicOperations["datetime=="](a, b); },
   now: () => new Date(),
-  today: () => { return moment().startOf("day").toDate(); },
+  today: () => {
+    const start = moment().startOf("day");
+    const y = start.year();
+    const m = start.month();
+    const d = start.date();
+    // tip: we use UTC to return same result as eg. new Date("2025-05-16")
+    const startUtc = moment.utc([y, m, d]);
+    return startUtc.toDate();
+  },
   start_of_today: () => { return moment().startOf("day").toDate(); },
   date_add: (date, val, dim) => { return moment(date).add(val, dim).toDate(); },
   datetime_add: (datetime, val, dim) => { return moment(datetime).add(val, dim).toDate(); },
   datetime_truncate: (datetime, dim) => { return moment(datetime).startOf(dim).toDate(); },
-  toLowerCase: (str) => str.toLowerCase(),
-  toUpperCase: (str) => str.toUpperCase(),
 };
 
 export function addRequiredJsonLogicOperations(jl = JL) {