+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Fix using incorrect date format in output to MongoDb (PR #1254)
   - Added date/datetime functions: `TODAY`, `START_OF_TODAY`, `RELATIVE_DATE`, `TRUNCATE_DATETIME` (PR #1256) (issue #1237)
   - Fix datetime functions: fix `NOW`, `RELATIVE_DATETIME` for MongoDb (PR #1256)
+  - Fixed comparing of dates with `==` and `!=` in JsonLogic by using new ops `date==`, `date!=`, `datetime==`, `datetime!=` (PR #1256)
 - 6.6.14
   - Fixed import from JsonLogic when like op is used inside group with some group op (PR #1225) (issue #1221)
   - Support React 19 (PR #1229) (issue #1205)